// src/client/api/generated/model/publishRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Spec Studio API
 * API local của Spec Studio. Lỗi chỉ trả mã (ErrorCode); FE tự dịch.
 * OpenAPI spec version: 1.0.0
 */
import type { PublishStepOutput } from './publishStepOutput';

export interface PublishRun {
  runId: string;
  workspaceId: string;
  file: string;
  steps: PublishStepOutput[];
  finished: boolean;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$ */
  createdAt: string;
  finishedAt: string | null;
}